Finished Admin Module HTML and TS
I finished the TS and HTML files for the Admin Module (user, team, and admin components are done).

I started work on the escalations module where I encountered the need for the getPopulatedTeam http request.  I wrote that request in the httpService and the updateComponent method on the escalations component.

Ready to start writing the escalations methods.